requirements.txt
streamlit
openai
google-generativeai
python-dotenv
bert-score
transformers
torch
pandas
tiktoken